dom/nodes/Element-childElementCount-svg.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="100%" height="100%" version="1.1" viewBox="0 0 400 400"><title>childElementCount</title><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><text x="200" y="40" fill="#000" font-size="25" text-anchor="middle">Test of childElementCount</text><text id="parentEl" x="200" y="70" fill="#000" font-size="20" text-anchor="middle">The result of <tspan id="first_element_child"><tspan>this</tspan> <tspan>test</tspan></tspan> is <tspan id="middle_element_child" font-weight="bold">unknown.</tspan> <tspan id="last_element_child" display="none">fnord</tspan></text><h:script><![CDATA[
test(function() {
  var parentEl = document.getElementById("parentEl")
  assert_true("childElementCount" in parentEl)
  assert_equals(parentEl.childElementCount, 3)
})
]]></h:script></svg>
